+++ src/Contracts/Compilation/Quota/AvailabilityInterface.php
@@ -41,12 +41,17 @@ use Teknoo\East\Paas\Compilation\Compiler\Exception\QuotasNotCompliantException;
  */
 interface AvailabilityInterface
 {
-    public function __construct(string $type, string $capacity, bool $isSoft);
+    public function __construct(string $type, string $capacity, string $require, bool $isSoft);
 
     /**
      * Return the current available remaining capacity for this quota
      */
     public function getCapacity(): string;
+
+    /**
+     * Return the current available remaining require capacity for this quota
+     */
+    public function getRequire(): string;
 
     /**
      * To update the quota from the deployment file to decrease it.