+++ tests/src/Compilation/Compiler/QuotaCompilerTest.php
@@ -53,12 +53,13 @@ class QuotaCompilerTest extends TestCase
             [
                 'category' => 'compute',
                 'type' => 'cpu',
-                'limit' => 2
+                'capacity' => 2,
+                'require' => 1,
             ],
             [
                 'category' => 'memory',
                 'type' => 'memory',
-                'limit' => '512Mi',
+                'capacity' => '512Mi',
             ]
         ];
     }